requirements.txt
dj-database-url==0.4.1
Django==1.9.9
gunicorn==19.6.0
virtualenv==15.0.3
whitenoise==3.2.1
psycopg2==2.5.4
djangorestframework==3.4.6
django-rest-auth==0.8.1
django-allauth==0.27.0
django-cors-headers=1.1.0